rewrite class Node for CheckStyle error again
