+++ src/managers/Node.java
@@ -1,0 +1,13 @@
+package managers;
+
+public class Node<Task> {
+    public Task task;
+    public Node<Task> next;
+    public Node<Task> prev;
+
+    public Node(Node<Task> prev, Task task, Node<Task> next) {
+        this.task = task;
+        this.next = null;
+        this.prev = null;
+    }
+}
